refactor: cleanup and error handling improvements
- Remove unused dependencies (Cypress, Jest, Babel)
- Remove redundant scripts in package.json
- Add test result folders to .gitignore
- Remove TwitchVideoGrid.vue component (duplicate)
- Add retry.ts utility function for attempt handling
- Integrate retry logic into video store
- Fix imports in SearchBar.vue and UserPreferences.vue
- Standardize @ alias usage
