--- frontend/cypress.config.js
@@ -1,22 +1,0 @@
-const { defineConfig } = require('cypress')
-
-module.exports = defineConfig({
-  e2e: {
-    baseUrl: 'http://localhost:5173',
-    supportFile: 'cypress/support/e2e.js',
-    specPattern: 'cypress/e2e/**/*.cy.{js,jsx,ts,tsx}',
-    viewportWidth: 1280,
-    viewportHeight: 720,
-    video: false,
-    screenshotOnRunFailure: true,
-    setupNodeEvents(on, config) {
-      // implement node event listeners here
-    },
-  },
-  component: {
-    devServer: {
-      framework: 'vue',
-      bundler: 'vite',
-    },
-  },
-})
