+++ frontend/package.json
@@ -11,10 +11,9 @@
     "test:unit:coverage": "vitest run --coverage",
     "test:e2e": "playwright test",
     "test": "npm run test:unit && npm run test:e2e",
-    "test:e2e:dev": "cypress open",
-    "test:coverage": "vitest run --coverage",
     "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs --fix --ignore-path .gitignore",
-    "format": "prettier --write src/"
+    "format": "prettier --write src/",
+    "clean": "rimraf node_modules/.tmp dist"
   },
   "dependencies": {
     "@headlessui/vue": "^1.7.19",
@@ -30,9 +29,6 @@
     "vue-toastification": "^2.0.0-rc.5"
   },
   "devDependencies": {
-    "@babel/core": "^7.26.10",
-    "@babel/preset-env": "^7.26.9",
-    "@jest/globals": "^29.7.0",
     "@pinia/testing": "^0.1.3",
     "@playwright/test": "^1.51.1",
     "@testing-library/vue": "^8.1.0",
@@ -41,18 +37,10 @@
     "@vitest/coverage-v8": "^3.0.9",
     "@vue/test-utils": "^2.4.6",
     "@vue/tsconfig": "^0.7.0",
-    "@vue/vue3-jest": "^29.2.6",
     "autoprefixer": "^10.4.19",
     "axios-mock-adapter": "^2.1.0",
-    "babel-jest": "^29.7.0",
-    "cypress": "^13.7.1",
     "eslint": "^8.57.0",
     "eslint-plugin-vue": "^9.24.0",
-    "expect": "^29.7.0",
-    "jest": "^29.7.0",
-    "jest-environment-jsdom": "^29.7.0",
-    "jest-transform-stub": "^2.0.0",
-    "jsdom": "^26.0.0",
     "postcss": "^8.4.38",
     "prettier": "^3.2.5",
     "tailwindcss": "^3.4.1",
